Multi threaded web server - fixes #23
